requirements.txt
google-cloud-aiplatform
hyperopt
pandas==1.5.3
numpy==1.24.3
kfp==2.5.0
scikit-learn==1.3.2
db-dtypes==1.2.0